:recycle: refactor theme, useTheme, & toggleTheme
