+++ client/@types/rebass.d.ts
@@ -1,0 +1,10 @@
+/* eslint-disable import/no-duplicates */
+declare module 'rebass/styled-components' {
+  import * as r from 'rebass';
+  import { CSSObject, CSSObject, StyledComponent } from 'styled-components';
+
+  export interface BaseProps {
+    sx?: CSSObject;
+  }
+  export = r;
+}
