--- client/@types/rebass/styled-components.d.ts
@@ -1,5 +1,0 @@
-// import { ButtonHTMLAttributes } from 'react';
-
-declare module 'rebass/styled-components';
-
-// export const Button: React.FC<ButtonHTMLAttributes>;
