+++ client/components/Select.tsx
@@ -1,20 +1,18 @@
 import * as React from 'react';
 import uuid from 'uuid/v4';
-
+import { Select } from '@rebass/forms';
 interface Props extends React.SelectHTMLAttributes<HTMLSelectElement> {
   options: string[];
 }
 
-const Select = ({ options, ...props }: Props): React.ReactElement<HTMLSelectElement> => {
+export default ({ options, ...props }: Props) => {
   return (
-    <select {...props}>
+    <Select {...props}>
       {options.map((o) => (
         <option key={uuid()} value={o}>
           {o}
         </option>
       ))}
-    </select>
+    </Select>
   );
 };
-
-export default Select;
