+++ client/hooks/useTheme.ts
@@ -1,13 +1,18 @@
+/* eslint-disable no-undef */
 /* eslint-disable no-nested-ternary */
 /* eslint-disable no-unused-expressions */
 import { useEffect, useState } from 'react';
-import { modes, getTheme } from '../utils/theme';
+import { getTheme, Mode, Theme } from '../utils/theme';
 
-const useTheme = () => {
-  const [theme, setTheme] = useState('light');
-  const [componentMounted, setComponentMounted] = useState(false);
+type ToggleTheme = () => void;
+type UseThemeValues = { theme: Theme; toggleTheme: ToggleTheme; componentMounted: boolean };
+type UseTheme = () => UseThemeValues;
 
-  const setMode = (mode: typeof modes[number]) => {
+const useTheme: UseTheme = () => {
+  const [theme, setTheme] = useState<Mode>('light');
+  const [componentMounted, setComponentMounted] = useState<boolean>(false);
+
+  const setMode = (mode: Mode) => {
     window.localStorage.setItem('theme', mode);
     setTheme(mode);
   };
@@ -21,7 +26,7 @@ const useTheme = () => {
   };
 
   useEffect(() => {
-    const localTheme = window.localStorage.getItem('theme');
+    const localTheme: Mode | undefined = (window as any & { theme: Mode }).localStorage.getItem('theme');
     window.matchMedia && window.matchMedia('(prefers-color-scheme: dark)').matches && !localTheme
       ? setMode('dark')
       : localTheme
