+++ client/hooks/window.d.ts
@@ -1,0 +1,5 @@
+declare global {
+  interface Window {
+    React: any;
+  }
+}
