+++ client/next-env.d.ts
@@ -1,2 +1,3 @@
+/* eslint-disable spaced-comment */
 /// <reference types="next" />
 /// <reference types="next/types/global" />
