+++ client/package.json
@@ -29,11 +29,7 @@
     "@apollo/react-hoc": "^3.1.3",
     "@apollo/react-hooks": "^3.1.3",
     "@graphql-codegen/typescript-react-apollo": "^1.9.1",
-    "@types/graphql": "^14.5.0",
-    "@types/lodash": "^4.14.149",
-    "@types/next": "^9.0.0",
-    "@types/react-dom": "^16.9.4",
-    "@types/styled-components": "^4.4.0",
+    "@rebass/forms": "^4.0.6",
     "@zeit/next-typescript": "^1.1.1",
     "apollo-boost": "^0.4.4",
     "apollo-cache-inmemory": "^1.6.3",
@@ -64,12 +60,18 @@
     "@graphql-codegen/cli": "^1.9.1",
     "@graphql-codegen/typescript": "^1.9.1",
     "@graphql-codegen/typescript-operations": "^1.9.1",
+    "@types/graphql": "^14.5.0",
+    "@types/lodash": "^4.14.149",
+    "@types/next": "^9.0.0",
     "@types/node": "^12.12.14",
     "@types/node-fetch": "^2.5.4",
     "@types/react": "^16.9.15",
+    "@types/react-dom": "^16.9.4",
     "@types/rebass": "^4.0.3",
+    "@types/styled-components": "^4.4.0",
     "@types/uuid": "^3.4.6",
     "@types/yup": "^0.26.26",
+    "@types/zeit__next-typescript": "^0.1.3",
     "@typescript-eslint/eslint-plugin": "^2.10.0",
     "@typescript-eslint/parser": "^2.10.0",
     "babel-eslint": "^10.0.3",
